Move chunks of code to separate files
